Refactor creating job application and contact, after creating job app… …lication -> create contact -> assign this contact to job application, add base class for stronlytypedid, add valueconverter for strongly type ids

+++ src/Modules/Users/YourCorporation.Modules.Users.Api/Entities/Role.cs
@@ -2,7 +2,7 @@
 
 namespace YourCorporation.Modules.Users.Api.Entities
 {
-    internal class Role : Entity<long>
+    internal class Role : Entity<int>
     {
         public string Name { get; set; }
